# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.AddFriend:
    properties:
      friendId:
        type: string
      groupId:
        type: integer
      lastReadTime:
        type: string
      userId:
        type: string
    type: object
...
  model.ChatGroup:
    properties:
      avatar:
        type: string
      createdAt:
        description: 创建时间
        type: string
      groupName:
        type: string
      id:
        description: 自增id
        type: integer
      intro:
        type: string
      notice:
        type: string
      updatedAt:
        description: 更新时间
        type: string
      userId:
        type: string
    type: object
  model.ChatGroupMember:
    properties:
      createdAt:
        description: 创建时间
        type: string
      groupId:
        type: integer
      id:
        description: 自增id
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      userId:
        type: string
      userInfo:
        $ref: '#/definitions/model.User'
    type: object
  model.ChatLog:
    properties:
      content:
        type: string
      createdAt:
        description: 创建时间
        type: string
      groupId:
        type: integer
      id:
        description: 自增id
        type: integer
      isDeleted:
        type: boolean
      isRevoked:
        type: boolean
      logType:
        type: integer
      msgType:
        type: integer
      operateBy:
        description: 记录管理端处置操作者，便于后续审计追溯
        type: string
      receiverId:
        type: string
      senderId:
        type: string
      updatedAt:
        description: 更新时间
        type: string
      userInfo:
        $ref: '#/definitions/model.User'
    type: object
  model.ChatLogQuery:
    properties:
      content:
        example: ""
        type: string
      groupId:
        type: integer
      page:
        example: 1
        type: integer
      pageSize:
        example: 20
        type: integer
      receiverId:
        type: string
      senderId:
        type: string
    type: object
...
  model.UpdateChatGroup:
    properties:
      avatar:
        description: 群头像（可选）
        type: string
      groupName:
        description: 群名称（可选）
        type: string
      intro:
        description: 群介绍（可选）
        type: string
      notice:
        description: 群公告（可选）
        type: string
    type: object
  model.UpdateReadTime:
    properties:
      groupId:
        type: integer
      receiverId:
        type: string
      senderId:
        type: string
    type: object
...
  model.User:
    properties:
      avatar:
        description: 用户头像
        example: https://******.com/aa.png
        type: string
      createdAt:
        description: 创建时间
        type: string
      email:
        description: 邮箱
        example: 123456789@qq.com
        type: string
      gender:
        description: 性别
        type: integer
      id:
        description: 自增id
        type: integer
      intro:
        description: 个人介绍
        example: 个人介绍
        type: string
      isAdmin:
        default: false
        description: 是否管理员
        type: boolean
      isLock:
        default: false
        description: 是否已锁
        type: boolean
      nickName:
        description: 用户昵称
        example: 酱
        type: string
      updatedAt:
        description: 更新时间
        type: string
      userId:
        description: 用户唯一id
        type: string
      userName:
        description: 用户名
        example: test
        maxLength: 12
        minLength: 4
        type: string
    required:
    - userName
    type: object
  response.ResType:
    properties:
      code:
        default: 0
        type: integer
      data: {}
      encrypt:
        type: string
      msg:
        default: 操作成功
        type: string
    type: object
host: localhost:9600
info:
  contact: {}
  description: zone服务端API服务
  title: Gin-Zone-Api
  version: "1.0"
paths:
...
  /mobile/chat/friends:
    get:
      consumes:
      - application/json
      description: 好友列表
      parameters:
      - description: User.ID
        in: query
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResType'
      security:
      - Authorization: []
      summary: 好友列表
      tags:
      - 聊天模块
    post:
      consumes:
      - application/json
      description: 添加好友或加入群聊
      parameters:
      - description: 需要上传的json
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.AddFriend'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AddFriend'
      security:
      - Authorization: []
      summary: 添加好友关系
      tags:
      - 聊天模块
  /mobile/chat/friends/{friendId}:
    delete:
      consumes:
      - application/json
      description: 删除好友
      parameters:
      - description: 好友id
        in: path
        name: friendId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResType'
      security:
      - Authorization: []
      summary: 删除好友
      tags:
      - 聊天模块
  /mobile/chat/groupMembers:
    get:
      consumes:
      - application/json
      description: 群成员列表
      parameters:
      - description: ChatGroup.ID
        in: query
        name: groupId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResType'
      security:
      - Authorization: []
      summary: 群成员列表
      tags:
      - 聊天模块
    post:
      consumes:
      - application/json
      description: 添加群成员
      parameters:
      - description: 需要上传的json
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.ChatGroupMember'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ChatGroupMember'
      security:
      - Authorization: []
      summary: 添加群成员
      tags:
      - 聊天模块
  /mobile/chat/groupMembers/{groupId}:
    delete:
      consumes:
      - application/json
      description: 退出群聊
      parameters:
      - description: 需要上传的json
        in: path
        name: groupId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResType'
      security:
      - Authorization: []
      summary: 退出群聊
      tags:
      - 聊天模块
  /mobile/chat/groupMembers/{groupId}/{userId}:
    delete:
      consumes:
      - application/json
      description: 群主/管理员删除群成员
      parameters:
      - description: 群组id
        in: path
        name: groupId
        required: true
        type: integer
      - description: 成员用户id
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResType'
      security:
      - Authorization: []
      summary: 删除群成员
      tags:
      - 聊天模块
  /mobile/chat/groups:
    get:
      consumes:
      - application/json
      description: 群组列表
      parameters:
      - description: User.ID
        in: query
        name: userId
        type: integer
      - description: Group.groupName
        in: query
        name: groupName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResType'
      security:
      - Authorization: []
      summary: 群组列表
      tags:
      - 聊天模块
    post:
      consumes:
      - application/json
      description: 创建群聊
      parameters:
      - description: 需要上传的json
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.ChatGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ChatGroup'
      security:
      - Authorization: []
      summary: 创建群聊
      tags:
      - 聊天模块
  /mobile/chat/groups/{groupId}:
    delete:
      consumes:
      - application/json
      description: 删除群聊
      parameters:
      - description: 群组id
        in: path
        name: groupId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResType'
      security:
      - Authorization: []
      summary: 删除群聊
      tags:
      - 聊天模块
    get:
      consumes:
      - application/json
      description: 获取群聊信息（群主/管理员/群成员可查看）
      parameters:
      - description: 群组id
        in: path
        name: groupId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResType'
      security:
      - Authorization: []
      summary: 群聊信息
      tags:
      - 聊天模块
    patch:
      consumes:
      - application/json
      description: 修改群聊信息（仅群主可修改）
      parameters:
      - description: 群组id
        in: path
        name: groupId
        required: true
        type: integer
      - description: 需要上传的json
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.UpdateChatGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResType'
      security:
      - Authorization: []
      summary: 修改群聊信息
      tags:
      - 聊天模块
  /mobile/chat/logs:
    post:
      consumes:
      - application/json
      description: 聊天记录
      parameters:
      - description: 需要上传的json
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/model.ChatLogQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ChatLog'
            type: array
      security:
      - Authorization: []
      summary: 聊天记录
      tags:
      - 聊天模块
  /mobile/chat/updateReadTime:
    post:
      consumes:
      - application/json
      description: 更新上次阅读信息时间
      parameters:
      - description: 需要上传的json
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/model.UpdateReadTime'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      security:
      - Authorization: []
      summary: 更新阅读时间
      tags:
      - 聊天模块
...
securityDefinitions:
  Authorization:
    description: jwt token 鉴权
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"